Move DisplayedFields to look_feel_tabs.__init__
